vector_store/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.4.0
langchain-openai>=0.2.0
langchain-google-genai>=1.0.0
langchain-mistralai>=0.2.0
openai>=1.0.0
faiss-cpu>=1.9.0
yandex-cloud-ml-sdk>=0.17.0
